Compare new videos to existing videos
Don’t just use the GUID cache, but also check the database’s `files`
table  to determine if we already have this video. Toward #56. Note
that there’s a major shortcoming here, which is that it identifies no
more than a single committee video per day, despite that there are
routinely more. We are not done here.