Temporarily disable deactivating legislators
Apparently due to a failed match on the House listing, it went and marked a whole bunch of legislators as about to leave office, including folks who were just elected, also including folks who long ago left office (Spruill, Putney). That's made a bit of a mess of things, because the temporary fix is to make all recently-updated legislators still in office...so Putney and Spruill are back at it, according to this listing.