Change how we handle un-insertable bills
It used to be that the only reason bills couldn't be added to the database (in practice) was because there was no record of the legislators. But now the legislature keeps emitting invalid CSV (e.g., using unescaped quotation marks), and that's the larger problem. So note that in the log file, but also don't give up on adding bills by a given legislator just because a prior bill failed.